Adds Bookmark and Comment Options to GameTree Node Menu.
Issue 940 - https://github.com/SabakiHQ/Sabaki/issues/940

When navigating the gametree, it is not obvious how to
bookmark or change colors of nodes.

To bookmark, you can use hotspots (ctrl+b), and to change color,
you can leave comments on the nodes.  But neither of these
options are available by right clicking a node, so unless you
happen to come across this functionality by leaving a comment
and happen to notice the impact, there is nothing that tells the
user that this functionality exists.

This commit rectifies this by adding these options when right
clicking a node in the gametree.